+++ testutils/src/main/java/com/google/android/exoplayer2/testutil/FakeTrackSelector.java
@@ -23,8 +23,8 @@ import com.google.android.exoplayer2.source.MediaSource.MediaPeriodId;
 import com.google.android.exoplayer2.source.TrackGroup;
 import com.google.android.exoplayer2.source.TrackGroupArray;
 import com.google.android.exoplayer2.trackselection.DefaultTrackSelector;
+import com.google.android.exoplayer2.trackselection.ExoTrackSelection;
 import com.google.android.exoplayer2.trackselection.MappingTrackSelector;
-import com.google.android.exoplayer2.trackselection.TrackSelection;
 import com.google.android.exoplayer2.upstream.BandwidthMeter;
 import java.util.ArrayList;
 import java.util.List;
@@ -41,8 +41,8 @@ public class FakeTrackSelector extends DefaultTrackSelector {
 
   /**
    * @param mayReuseTrackSelection Whether this {@link FakeTrackSelector} will reuse {@link
-   *     TrackSelection}s during track selection, when it finds previously-selected track selection
-   *     using the same {@link TrackGroup}.
+   *     ExoTrackSelection}s during track selection, when it finds previously-selected track
+   *     selection using the same {@link TrackGroup}.
    */
   public FakeTrackSelector(boolean mayReuseTrackSelection) {
     this(new FakeTrackSelectionFactory(mayReuseTrackSelection));
@@ -54,18 +54,18 @@ public class FakeTrackSelector extends DefaultTrackSelector {
   }
 
   @Override
-  protected TrackSelection.@NullableType Definition[] selectAllTracks(
+  protected ExoTrackSelection.@NullableType Definition[] selectAllTracks(
       MappedTrackInfo mappedTrackInfo,
       @Capabilities int[][][] rendererFormatSupports,
       @AdaptiveSupport int[] rendererMixedMimeTypeAdaptationSupports,
       Parameters params) {
     int rendererCount = mappedTrackInfo.getRendererCount();
-    TrackSelection.@NullableType Definition[] definitions =
-        new TrackSelection.Definition[rendererCount];
+    ExoTrackSelection.@NullableType Definition[] definitions =
+        new ExoTrackSelection.Definition[rendererCount];
     for (int i = 0; i < rendererCount; i++) {
       TrackGroupArray trackGroupArray = mappedTrackInfo.getTrackGroups(i);
       boolean hasTracks = trackGroupArray.length > 0;
-      definitions[i] = hasTracks ? new TrackSelection.Definition(trackGroupArray.get(0)) : null;
+      definitions[i] = hasTracks ? new ExoTrackSelection.Definition(trackGroupArray.get(0)) : null;
     }
     return definitions;
   }
@@ -75,7 +75,7 @@ public class FakeTrackSelector extends DefaultTrackSelector {
     return fakeTrackSelectionFactory.trackSelections;
   }
 
-  private static class FakeTrackSelectionFactory implements TrackSelection.Factory {
+  private static class FakeTrackSelectionFactory implements ExoTrackSelection.Factory {
 
     private final List<FakeTrackSelection> trackSelections;
     private final boolean mayReuseTrackSelection;
@@ -86,14 +86,14 @@ public class FakeTrackSelector extends DefaultTrackSelector {
     }
 
     @Override
-    public TrackSelection[] createTrackSelections(
-        TrackSelection.@NullableType Definition[] definitions,
+    public ExoTrackSelection[] createTrackSelections(
+        ExoTrackSelection.@NullableType Definition[] definitions,
         BandwidthMeter bandwidthMeter,
         MediaPeriodId mediaPeriodId,
         Timeline timeline) {
-      TrackSelection[] selections = new TrackSelection[definitions.length];
+      ExoTrackSelection[] selections = new ExoTrackSelection[definitions.length];
       for (int i = 0; i < definitions.length; i++) {
-        TrackSelection.Definition definition = definitions[i];
+        ExoTrackSelection.Definition definition = definitions[i];
         if (definition != null) {
           selections[i] = createTrackSelection(definition.group);
         }
@@ -101,7 +101,7 @@ public class FakeTrackSelector extends DefaultTrackSelector {
       return selections;
     }
 
-    private TrackSelection createTrackSelection(TrackGroup trackGroup) {
+    private ExoTrackSelection createTrackSelection(TrackGroup trackGroup) {
       if (mayReuseTrackSelection) {
         for (FakeTrackSelection trackSelection : trackSelections) {
           if (trackSelection.getTrackGroup().equals(trackGroup)) {